MBS-8242: Link to specific relationship type docs from rel edits
This adds a link to the doc page(s) for the relationship type(s)
involved in a relationship edit (add, edit or remove).
It seems very useful to have direct links to the documentation
(often including specific guidelines)
directly in the place people need to vote on the changes from.

We pass an array to allow multiple types in EditRelationship;
we pass the rels through a Set to remove duplicates
when EditRelationship does not change the rel type.